DAL entity improvements:
- Improve id type constitency
- Move entity constants to dedicated class
- Move role name to dedictaed class
- Remove DbSet properties from context
- Split Entity configuration by classes and move to fluent api
- Some additional refactoring

+++ ForumTask.DAL/DependencyInjection/DalServiceCollectionExtensions.cs
@@ -17,8 +17,8 @@ namespace ForumTask.DAL.DependencyInjection
 
             services.AddScoped(typeof(IRepository<>), typeof(GenericRepository<>));
 
-            services.AddIdentity<User, IdentityRole<int>>(ConfigureIdentity)
-                .AddRoles<IdentityRole<int>>()
+            services.AddIdentity<User, IdentityRole<long>>(ConfigureIdentity)
+                .AddRoles<IdentityRole<long>>()
                 .AddEntityFrameworkStores<ForumContext>();
         }
